[#40] chore: Move SurveyOutro into surveys directory
